Move WidgetPod into core.rs
